Use 'uint32' instead of 'int64' for blob sizes everywhere.
Not just in blob.SizedRef, but in blobserver.Fetch and
blobserver.FetchStreaming, too.
Blobs have a max size of 10-32 MB anyway, and the index.Corpus is now using
uint32 to save memory.

+++ cmd/camput/cache.go
@@ -25,8 +25,8 @@ import (
 
 // A HaveCache tracks whether a remove blobserver has a blob or not.
 type HaveCache interface {
-	StatBlobCache(br blob.Ref) (size int64, ok bool)
-	NoteBlobExists(br blob.Ref, size int64)
+	StatBlobCache(br blob.Ref) (size uint32, ok bool)
+	NoteBlobExists(br blob.Ref, size uint32)
 	Close() error
 }
